Move extensions for Conduction
