+++ Conduction/ITransmission.cs
@@ -9,10 +9,6 @@ namespace YggdrAshill.Nuadha.Conduction
     /// <typeparam name="TSignal">
     /// Type of <see cref="ISignal"/> to transmit.
     /// </typeparam>
-    /// <remarks>
-    /// Should be generated by <see cref="IPropagation{TSignal}"/> combined with <see cref="IGeneration{TSignal}"/>.
-    /// In detail, please see <see cref="PropagationExtension.Transmit{TSignal}(IPropagation{TSignal}, IGeneration{TSignal})"/>.
-    /// </remarks>
     public interface ITransmission<TSignal> :
         IProduction<TSignal>,
         IEmission,
