+++ YggdrAshill.Nuadha.Specification/Conduction/ConductionExtensionSpecification.cs
@@ -1,5 +1,6 @@
 using NUnit.Framework;
 using YggdrAshill.Nuadha.Signalization;
+using YggdrAshill.Nuadha.Conduction;
 using System;
 
 namespace YggdrAshill.Nuadha.Specification
@@ -9,6 +10,8 @@ namespace YggdrAshill.Nuadha.Specification
     {
         private PropagateSignal propagation;
 
+        private GenerateSignal generation;
+
         private ConsumeSignal consumption;
 
         [SetUp]
@@ -16,11 +19,27 @@ namespace YggdrAshill.Nuadha.Specification
         {
             propagation = new PropagateSignal();
 
+            generation = new GenerateSignal();
+
             consumption = new ConsumeSignal();
         }
 
         [Test]
-        public void ShouldTransmitSignal()
+        public void ShouldTransmitSignalWithGeneration()
+        {
+            var transmission = propagation.Transmit(generation);
+
+            var cancellation = transmission.Produce(consumption);
+
+            transmission.Emit();
+
+            Assert.AreEqual(generation.Generated, consumption.Consumed);
+
+            cancellation.Cancel();
+        }
+
+        [Test]
+        public void ShouldTransmitSignalWithFunction()
         {
             var expected = new Signal();
             var transmission = propagation.Transmit(() =>
@@ -40,6 +59,17 @@ namespace YggdrAshill.Nuadha.Specification
         [Test]
         public void CannotTransmitWithNull()
         {
+            Assert.Throws<ArgumentNullException>(() =>
+            {
+                var transmission = default(IPropagation<Signal>).Transmit(generation);
+
+            });
+
+            Assert.Throws<ArgumentNullException>(() =>
+            {
+                var transmission = propagation.Transmit(default(IGeneration<Signal>));
+            });
+
             Assert.Throws<ArgumentNullException>(() =>
             {
                 var transmission = default(IPropagation<Signal>).Transmit(() => new Signal());
